chore(#821): add test for install / restructure bridge test
