+++ client/tests/e2e/base.py
@@ -1,13 +1,29 @@
+from time import sleep
+from typing import List
+import docker
+import docker.errors
+from kubernetes.config import load_kube_config
+
+import requests
+
+from kubernetes.client import (
+    CoreV1Api,
+    RbacAuthorizationV1Api,
+    AppsV1Api,
+    CustomObjectsApi,
+)
+from pytest_kubernetes.providers import AClusterManager
+import pytest
+from unittest import TestCase
+
+from click.testing import CliRunner, Result
 from copy import deepcopy
 import os
-from time import sleep
 
 from click import BadParameter
-from click.testing import CliRunner
 from docker.context import ContextAPI
 from gefyra.cli.utils import check_connection_name
 
-import pytest
 
 from gefyra.api.clients import list_client, write_client_file
 from gefyra.api.install import LB_PRESETS
@@ -784,3 +800,81 @@ class GefyraBaseTest(GefyraTestMixin):
                 config=default_configuration, name="foo", namespace="default"
             )
         self.assertIn("Pod foo not found.", str(rte.exception))
+
+
+LOCAL_CONTAINER_NAME = "gefyra-new-backend"
+
+
+@pytest.mark.usefixtures("operator")
+@pytest.mark.usefixtures("tmp_path")
+@pytest.mark.usefixtures("demo_backend_image")
+class GefyraTestCase(TestCase):
+    provider = "k3d"
+
+    @pytest.fixture(autouse=True)
+    def _capture_request(self, request):
+        self._request = request
+
+    def _init_docker(self):
+        self.DOCKER_API = docker.from_env()
+
+    def _init_kube_api(self):
+        load_kube_config(self.kubeconfig)
+        self.K8S_CORE_API = CoreV1Api()
+        self.K8S_RBAC_API = RbacAuthorizationV1Api()
+        self.K8S_APP_API = AppsV1Api()
+        self.K8S_CUSTOM_OBJECT_API = CustomObjectsApi()
+
+    def setUp(self):
+        self.operator: AClusterManager = self._request.getfixturevalue("operator")
+        self.tmp_path = self._request.getfixturevalue("tmp_path")
+        self.demo_backend_image = self._request.getfixturevalue("demo_backend_image")
+        self.kubeconfig = str(self.operator.kubeconfig)
+        self._init_kube_api()
+        self._init_docker()
+        self.operator.apply("tests/fixtures/nginx_exposed.yaml")
+
+    def tearDown(self):
+        containers = ["gefyra-cargo-pytest-gefyra", LOCAL_CONTAINER_NAME]
+        for container in containers:
+            try:
+                self.DOCKER_API.containers.get(container).remove(force=True)
+            except docker.errors.NotFound:
+                pass
+        return super().tearDown()
+
+    def cmd(self, command: str, params: List[str]) -> Result:
+        load_kube_config(self.kubeconfig)
+        from gefyra.cli.main import cli
+
+        runner = CliRunner()
+        res = runner.invoke(
+            cli,
+            ["--kubeconfig", str(self.operator.kubeconfig), command, *params],
+            catch_exceptions=True,
+        )
+        if res.exit_code != 0:
+            raise AssertionError(
+                f"Command failed: {res.output}\nExit code: " + str(res.exit_code)
+            )
+        return res
+
+    def assert_get_contains(
+        self, url: str, expected_content: str, retries: int = 10, headers: dict = None
+    ):
+        """
+        Helper function to assert that a GET request to a URL contains expected content.
+        Retries the request if the content is not found.
+        """
+        while retries > 0:
+            try:
+                response = requests.get(url, headers=headers)
+            except Exception:
+                continue
+            if expected_content in response.text:
+                return
+            retries -= 1
+            sleep(1)
+        raise AssertionError(
+            f"Expected content '{expected_content}' not found in response from {url}."
+        )
